Crate spirit_cfg_helpers

Spirit configuration helpers

These extensions are meant to integrate into the spirit configuration framework. They aim at making the user experience around application's configuration more smooth. Specifically, they allow dumping the configuration collected through all the config files and environment variables and printing help about all the configuration options the application accepts.

Features

By default, all features are turned on. However, it is possible to opt out of some to cut down on dependencies. Specifically:

toml and json features enable dumping in the respective formats.

cfg-help enables the printing of configuration help.

Structs

CfgDump
A command line fragment to add --dump-config to allow showing loaded configuration.

CfgHelp
A command line options fragment to add the --help-config option.

Opts
A combination of the CfgDump and CfgHelp fragments.

Functions

cfg_store
An extension to store configuration to some global-ish storage.

config_logging
An extension to log changes of configuration every time it is reloaded.